requirements.txt
aiogram==3.4.1
aiofiles==23.2.1
aiosqlite==0.19.0
httpx==0.27.0
APScheduler==3.10.4
# Pin pydantic in the aiogram-supported range (has wheels for 3.13)
pydantic==2.5.3
python-dateutil==2.8.2
asyncio-throttle==1.0.2
python-dotenv==1.0.1